+++ domain/src/main/java/tachiyomi/domain/manga/model/MangaWithChapterCount.kt
@@ -1,0 +1,6 @@
+package tachiyomi.domain.manga.model
+
+data class MangaWithChapterCount(
+    val manga: Manga,
+    val chapterCount: Long,
+)